Change int8 to intN for tinyid/shortid

+++ spidermonkey/jsscopeinlines.h
@@ -214,7 +214,7 @@ inline
 JSScopeProperty::JSScopeProperty(jsid id, JSPropertyOp getter, JSPropertyOp setter,
                                  uint32 slot, uintN attrs, uintN flags, intN shortid)
   : id(id), rawGetter(getter), rawSetter(setter), slot(slot), attrs(uint8(attrs)),
-    flags(uint8(flags)), shortid(int16(shortid))
+    flags(uint8(flags)), shortid(int32(shortid))
 {
     JS_ASSERT_IF(getter && (attrs & JSPROP_GETTER), getterObj->isCallable());
     JS_ASSERT_IF(setter && (attrs & JSPROP_SETTER), setterObj->isCallable());
